requirements.txt
numpy>=1.24.0
opencv-python==4.11.0.86
mediapipe==0.10.21
pydantic==2.11.3
openai==1.76.0
tiktoken==0.9.0
yt-dlp>=2025.4.30
assemblyai==0.40.2
PyYAML>=6.0.0
Jinja2==3.1.5
streamlit==1.41.1
dotenv==0.9.9